restruct

+++ src/modules/config_directory.py
@@ -3,12 +3,12 @@ import sys
 import json
 import shutil
 
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
-from variables import (
+from vars import (
     _PROJECT_CONFIG_PATH, 
     _SETTINGS_FILE_PATH, 
-    settings_content
-    )
+    settings_content,
+    debug
+)
 
 def configDirectory():
     if not os.path.exists(_PROJECT_CONFIG_PATH):
@@ -18,7 +18,7 @@ def configDirectory():
         with open(_SETTINGS_FILE_PATH, "w") as f:
             json.dump(settings_content, f, indent=2)
 
-if __name__ == "__main__":
+if __name__ == "__main__" and debug:
     if os.path.exists(_PROJECT_CONFIG_PATH):
         shutil.rmtree(_PROJECT_CONFIG_PATH)
